requirements.txt
python-telegram-bot==20.0
pytube==12.0.1
TikTokApi==0.10.0
instaloader==4.5.8
